Solution: Day 2
A bad run, the server lagged temporarily and I then used an invalid link
for some time which cost a few minutes. I also (re-)defined a variable inside
a loop, why can't python have scoped variables??? The second part went
awful for the simple reason that I could not come up with a simple strategy
and rewrote it 3 times and in that process messed up the logic borrowed
from part 1. I allways got the same answer as the test-input as well which
misled me for a while. Next day will be better.