Solution: Day 3
Today was a good day, I funnily enough predicted the solution would use
regex and had pre-opened `findall` in the documentation. Didn't need it
but still I called it! The drop in score was due to me forgetting that
*don't* is not spelled *dont'*... Also that the `mul` instruction starts
out as active. But still, best score I've gotten **so far**!

Yesterday's bad score is all in the past.