Solution: Day 4
As is evident by the amount of code, I could not come up with a way of
getting the diagonals effectively in part 1. Maybe a better strategy
would have been to do more like part 2 in part 1, going trough each
coordinate and expanding the search from that point. At least I reversed
the trend of dropping in ranking from doing part 2, but it's pretty
clear why...