Solution: Day 6
I don't think it could have gone worse. That example input was
basically ADVERSARIAL, nothing about it related to my input and I was
too shortsighted to look beyond it. I think it's best at this point to
abandon the idea of reaching top 1000. Unless this was an outlier, I
don't have the skills required.