Solution: Day 9
I think I made it harder for myself in the first part, thought I was
doing something clever but of course hade to pay the time cost of
implementing it and it wasnt *that* fast either. And since part 2
couldn't work with that solution I had to rewrite it and while doing to
messed up which index I was inserting spaces into which led me down a
lengthy debugging jurney. It would be fun trying to find an effective
solution for part 2.