Solution: Day 11
I don't know why I'm so mad about todays ranking. I think it is because
I knew the solution right away but I wrote `10 % (e - 1)` instead of
`10 % (e // 2)` to split the numbers and had to whip out a repl just to
realize something so **simple**. But I've played AoC enough to know the
oldest trick Eric has in his book which is why I jumped around 5000 from
part 1 to part 2 by writing the algorithm so that it would anticipate
part 2.