Solution: Day 14
Fun with a unique problem, should have picked up quicker on the
"congregation" every 103 seconds after x seconds but I thought it was a
pointless artifact. Given the first occurance was more than 2 hours in,
it was a good thing that after I "only" 10 minutes gave up on watching
everything at 2x speed and looked for patterns.

It was also funny to look at the leaderboards today, easy to pick out the
proompt "engineers" as they were under half a minute on part one but then
having no placement on part 2. It's almost like they're not capable of
image recognition...