Solution: Day 15
Today was a nice problem, I was fearing it would be one of those
"now repeat the moves until the boxes return to a previous position" so
I was glad that it was just a more involved algorithm. I thought I was
doing really bad on time but even the top 100 leaderboard struggled for
over 15 minutes so it seems it was just a problem without shortcuts. I
want more problems like these for the next few days, it's nice knowing
that it's more about problem solving than being a fast typer.