Solution: Day 16
This was the worse day yet, part 1 I landed a leaderboard position of
around 400 but then I for part 2 I could not for the life of me figure
out how to do it without brute forcing. There is some serious gap of
knowledge in my understanding of shortest path algorithms.

The brute force solution in the end is just looking back at each turn
I've made and redoing the calculation from there, but this ended up with
a massive amount of trial and error that took around 30s the first time
I got it fast enough.