Solution: Day 17
Good things come to those that wait, today was GREAT. Classic Eric move
to do a Synacore esque challange where you have to dissassemble the
input, f***ing love those. I could have gotten an even better placement
had I realized that the example input was a red herring, but it was good
to use it to realize that ANY output could be achieved since the program
is a decoder, it was just that hade the solution for around 15 minutes
before I tried to submit an answer. Hopefully I'm not breaking rules by
committing my dissassembly, since you could theoretically retrieve the
input I had, with one placement off.