Solution: Day 19
Nice and easy problem, messed up a bit on part one by not using a
visited set for each trial. It was hard (for me) to detect the error
as well as I thought the error was that `startswith` had some extra
functionality that I wasn't taking into account when comparing
strings. But it was simply that I was brute forcing too hard. I liked
the dynamic programming approach in part 2, nice and simple.