Solution: Day 20
Phew, this one kicked my ass. Didn't finish it in time before leaving
for school so had to do pick it up once I got back. I first tried
exploring the maze with added rules to handle passing through walls but
I ended up with some wierd off-by-one error. Solution was finding pairs of
points some distance from both the start and finish and let the
manhattan distance determine if there exists a time save by cheating.
There are probably some nice criteria that would reduce the amount of
pairs of points that need to be considered, but the the current solution
"only" takes around half a minute so I'm not gonna bother improving it.