Solution: Day 21
Amazing. Loved todays challange, it took some time to figure out aspects
like how it's allways best to step many steps in the same direction
consecutively as well as the fact that every step had to press the A
button means that every subsequent robot need only to look at inputs
leading up to a A-button press. This led me to the dynamic programming
solution for part 2.

I also ran into one of the worst "bugs" I've ever had to deal with;
due to it not being very unlikely that any of the shortest paths at any
stage is also a component of the total solution, I made the mistake to
assume after a SINGLE test that I only needed to look at one of the
shortest paths at any layer. However, this is only true for the last
layer. Tragically, sometimes the seed used for the hash-set ordered one
of the correct paths of the keypad first when picking the shortest, and
this led to me making the first incorrect assumption and also messed
with my head when trying out different resolutions. Although I picked up
on the random-ness of it and correctly assumed it had to do with
hashing, I made the second (major-)blunder of thinking it was a
hash-collision in the cashed recursion-function. It was a huge relief to
finally detirmine that the hash-set being used for the keypad-layer was
the culprit.

Again, amazing puzzle.