Solution: Day 22
The change in the PRICE, not the secret number... Oh well that's what I
get for not reading the problem thoroughly. Other than that, nice and
simple problem, was easy to solve with the right datastruct. All that's
left to see is how fast I can get it to run in rust...