Solution: Day 23
This should have been easy, but I just could'nt figure out the algorithm
for the last part for the longest time. I knew it was something close to
what I had in the end, but I just kept making mistakes and rewrote it
completely 3 or 4 times. I also didn't read the problem statment properly
and solved for interconnections instead of direct connections. And guess
what? I made that mistake for BOTH parts...

I'm not as certain anymore with graphs it seems, but if there are any
more graph-problems, at least I'll be warmed up!