Solution: Day 24
Every year there is allways that one problem which is both feasable and
easier to do manually and I think today hosted that problem. Together
with the fact that it's Christmas Eve, I've decided to not write a
general solution today.

To get my solution, I looked at a rendering of the graph looked for
irregularities *in the shape of the graph* that I used to determine some
regular patterns. Then I just wrote a simple program to output any
deviation from these patterns and manually checked for a possible swaps.

However, I do think a general solution would be outlined as; use the
regular structure for both `zxx` and the OR operation as an expected
pattern and with a DFT output all errors that exist and associate them
to the OR operation or the z-output. Find overlap of these errors and
hopefully it will even out to 4 swaps! And then some edge cases...