Solution: Day 25
Thats all for this year! Not much to say about today, it's fast enough
to just "brute force".

I'm happy with my effort in this year's AoC! Although I didn't reach my
goal of averaging sub 1000 (looking at days I did at launch or had the
time in the morning to do; on average 1600) I am still happy that I was
able to do each challenge. I'm also happy about these logs I've been
writing throughout, they have been a great source of reflection and
helped me to get over not getting the ranking I was aiming for and
to rather focus on the learning experience. I found while reading them
back that I had a tendency to blame a bad solution or ranking on
"silly" mistakes I made rather than taking ownership of them. Making
mistakes are part of your (lack of) skill and shouldn't be observed as
an exception.

I REALLY hope this wasn't the last entry in AoC. The reason I have my
doubts about this tradition is the abundance of llm-solutions in the
leaderboard (how many is up for debate, but some accounts hade public
repo's showing system prompts for AoC) and the way that it spoils the
competition (between humans). I would be discuraged to have a creation
I've made be used as a benchmarking tool so I'm hoping Eric still has
the motivation to keep doing this, next year I'll have my sub 1000!